# Conflicts:
#	changelog.md
#	common/src/main/java/net/mca/client/gui/SkinLibraryScreen.java
#	common/src/main/java/net/mca/mixin/client/MixinTranslationStorage.java

+++ common/src/main/java/net/mca/server/world/data/Village.java
@@ -90,7 +90,7 @@ public class Village implements Iterable<Building> {
             populationThreshold = v.getFloat("populationThresholdFloat");
         }
         if (v.contains("marriageThresholdFloat")) {
-            marriageThreshold = v.getInt("marriageThresholdFloat");
+            marriageThreshold = v.getFloat("marriageThresholdFloat");
         }
         this.world = world;
